refactor(KERN-1560): move extractTableName to shared utils/ast-table-ref
Generic utility for resolving the primary table name from a SelectNode's
FROM clause. Reusable by any module working with DuckDB ASTs.

work-item: KERN-1560
